requirements.txt
aiohttp~=3.12.13
aiosqlite~=0.21.0
discord.py==2.5.2
python-dotenv~=1.1.1
requests~=2.32.4
table2ascii~=1.1.3